Create shell script to launch the prime and probe and then launch the rattle.
This script first launches the L1-capture. Then while the capture is running it
will keep launching rattles.

Note if the number of samples passed to captures is too high and/or the sleep
delay too low then capture may not get scheduled enough time to complete in
a reasonable timeframe